Validate contextDataFactory export type at startup
