+++ packages/nuxt/deno.json
@@ -21,7 +21,7 @@
     ]
   },
   "tasks": {
-    "test": "deno test",
+    "test": "deno test --allow-env",
     "check": "deno fmt --check && deno lint && deno check src/*.ts src/runtime/server/*.ts"
   }
 }
